fix: write sink info and volume on any event
This is a bit of a hack.

When a device is removed, PulseAudio continues to
send events about the old sink, which overwrites
the data about the fresh sink.

To prevent this, on any event, I loop through all
the available sinks and select the newest one.
This causes multiple file writes in a short span
of time.

A better way of handling this would be to provide
information about all sinks, in addition to
the index of the newest sink.